+++ src/Draco.Compiler.Tests/Semantics/SemanticModelTests.cs
@@ -1,0 +1,104 @@
+using System.Collections.Immutable;
+using Draco.Compiler.Api;
+using Draco.Compiler.Api.Syntax;
+using Draco.Compiler.Internal.Binding;
+using Draco.Compiler.Internal.Symbols;
+using Draco.Compiler.Internal.Types;
+using static Draco.Compiler.Api.Syntax.SyntaxFactory;
+
+namespace Draco.Compiler.Tests.Semantics;
+
+public sealed class SemanticModelTests : SemanticTestsBase
+{
+    // Reported in https://github.com/Draco-lang/Compiler/issues/220
+    [Fact]
+    public void RequestingDiagnosticsAfterSymbolDoesNotDiscardDiagnostic()
+    {
+        // func main() {
+        //     var x;
+        // }
+
+        // Arrange
+        var tree = SyntaxTree.Create(CompilationUnit(FunctionDeclaration(
+            "main",
+            ParameterList(),
+            null,
+            BlockFunctionBody(
+                DeclarationStatement(VariableDeclaration("x"))))));
+
+        var xDecl = tree.FindInChildren<VariableDeclarationSyntax>(0);
+
+        // Act
+        var compilation = Compilation.Create(ImmutableArray.Create(tree));
+        var semanticModel = compilation.GetSemanticModel(tree);
+
+        var xSym = GetInternalSymbol<LocalSymbol>(semanticModel.GetDefinedSymbol(xDecl));
+        var diags = semanticModel.Diagnostics;
+
+        // Assert
+        Assert.Single(diags);
+        AssertDiagnostic(diags, TypeCheckingErrors.CouldNotInferType);
+        Assert.True(xSym.Type.IsError);
+    }
+
+    [Fact]
+    public void RequestingSymbolAfterDiagnosticsDoesNotDiscardDiagnostic()
+    {
+        // func main() {
+        //     var x;
+        // }
+
+        // Arrange
+        var tree = SyntaxTree.Create(CompilationUnit(FunctionDeclaration(
+            "main",
+            ParameterList(),
+            null,
+            BlockFunctionBody(
+                DeclarationStatement(VariableDeclaration("x"))))));
+
+        var xDecl = tree.FindInChildren<VariableDeclarationSyntax>(0);
+
+        // Act
+        var compilation = Compilation.Create(ImmutableArray.Create(tree));
+        var semanticModel = compilation.GetSemanticModel(tree);
+
+        var diags = semanticModel.Diagnostics;
+        var xSym = GetInternalSymbol<LocalSymbol>(semanticModel.GetDefinedSymbol(xDecl));
+
+        // Assert
+        Assert.Single(diags);
+        AssertDiagnostic(diags, TypeCheckingErrors.CouldNotInferType);
+        Assert.True(xSym.Type.IsError);
+    }
+
+    [Fact]
+    public void RequestingSymbolMultipleTimesDoesNotMultiplyDiagnostic()
+    {
+        // func main() {
+        //     var x;
+        // }
+
+        // Arrange
+        var tree = SyntaxTree.Create(CompilationUnit(FunctionDeclaration(
+            "main",
+            ParameterList(),
+            null,
+            BlockFunctionBody(
+                DeclarationStatement(VariableDeclaration("x"))))));
+
+        var xDecl = tree.FindInChildren<VariableDeclarationSyntax>(0);
+
+        // Act
+        var compilation = Compilation.Create(ImmutableArray.Create(tree));
+        var semanticModel = compilation.GetSemanticModel(tree);
+
+        _ = GetInternalSymbol<LocalSymbol>(semanticModel.GetDefinedSymbol(xDecl));
+        _ = GetInternalSymbol<LocalSymbol>(semanticModel.GetDefinedSymbol(xDecl));
+        _ = GetInternalSymbol<LocalSymbol>(semanticModel.GetDefinedSymbol(xDecl));
+        var diags = semanticModel.Diagnostics;
+
+        // Assert
+        Assert.Single(diags);
+        AssertDiagnostic(diags, TypeCheckingErrors.CouldNotInferType);
+    }
+}
