+++ src/Draco.Compiler.Tests/Semantics/TypeCheckingTests.cs
@@ -155,36 +155,6 @@ public sealed class TypeCheckingTests : SemanticTestsBase
     }
 
     [Fact]
-    public void GatheringDiagnostics()
-    {
-        // func main() {
-        //     var x;
-        // }
-
-        // Arrange
-        var tree = SyntaxTree.Create(CompilationUnit(FunctionDeclaration(
-            "main",
-            ParameterList(),
-            null,
-            BlockFunctionBody(
-                DeclarationStatement(VariableDeclaration("x"))))));
-
-        var xDecl = tree.FindInChildren<VariableDeclarationSyntax>(0);
-
-        // Act
-        var compilation = Compilation.Create(ImmutableArray.Create(tree));
-        var semanticModel = compilation.GetSemanticModel(tree);
-
-        var xSym = GetInternalSymbol<LocalSymbol>(semanticModel.GetDefinedSymbol(xDecl));
-        var diags = semanticModel.Diagnostics;
-
-        // Assert
-        Assert.Single(diags);
-        AssertDiagnostic(diags, TypeCheckingErrors.CouldNotInferType);
-        Assert.True(xSym.Type.IsError);
-    }
-
-    [Fact]
     public void LocalVariableIncompatibleType()
     {
         // func main() {
